requirements.txt
pandas
psycopg2-binary
pyarrow
flake8
pytest
python-dotenv